+++ src/city/repository/CityRepository.ts
@@ -1,0 +1,5 @@
+import { City } from 'src/city/model/City';
+
+export interface CityRepository {
+  getByDepartment(departmentId: number): Promise<City[]>;
+}
