+++ src/city/repository/InMemoryCityRepository.ts
@@ -1,0 +1,16 @@
+import { City } from 'src/city/model/City';
+import { CityRepository } from './CityRepository';
+
+export class InMemoryCityRepository implements CityRepository {
+  getByDepartment(departmentId: number): Promise<City[]> {
+    const list: City[] = [];
+
+    list.push(new City(5, 9, 2, 'Medellín'));
+
+    const response: City[] = list.filter((city) => {
+      return city.departmentId === departmentId;
+    });
+
+    return Promise.resolve(response);
+  }
+}
