+++ src/department/repository/DepartmentRepository.ts
@@ -1,0 +1,5 @@
+import { Department } from 'src/department/model/Department';
+
+export interface DepartmentRepository {
+  getAll(): Promise<Department[]>;
+}
