--- src/repository/CityRepository.ts
@@ -1,5 +1,0 @@
-import { City } from 'src/model/City';
-
-export interface CityRepository {
-  getByDepartment(departmentId: number): Promise<City[]>;
-}
