--- src/repository/DepartmentRepository.ts
@@ -1,5 +1,0 @@
-import { Department } from 'src/model/Department';
-
-export interface DepartmentRepository {
-  getAll(): Promise<Department[]>;
-}
