--- src/repository/InMemorySchoolRepository.ts
@@ -1,15 +1,0 @@
-import { School } from 'src/model/School';
-import { SchoolRepository } from './SchoolRepository';
-
-export class InMemorySchoolRepository implements SchoolRepository {
-  getByCity(cityId: number): Promise<School[]> {
-    const list: School[] = [];
-
-    list.push(new School(1, 3, 'Colegio Teresiano'));
-
-    const response: School[] = list.filter((school) => {
-      return school.cityId === cityId;
-    });
-    return Promise.resolve(response);
-  }
-}
