--- src/repository/SchoolRepository.ts
@@ -1,5 +1,0 @@
-import { School } from 'src/model/School';
-
-export interface SchoolRepository {
-  getByCity(cityId: number): Promise<School[]>;
-}
