+++ src/school/model/School.ts
