+++ src/school/repository/SchoolRepository.ts
@@ -1,0 +1,5 @@
+import { School } from 'src/school/model/School';
+
+export interface SchoolRepository {
+  getByCity(cityId: number): Promise<School[]>;
+}
